html/semantics/embedded-content/the-iframe-element/iframe_javascript_url_01.htm

<!DOCTYPE html><html><head><title>javascript: URL creating a document in an about:blank iframe</title>
<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/embedded-content-0.html#the-iframe-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log">FAILED (test did not run)</div>

<iframe src="about:blank" name="ifr1"></iframe>
<iframe src="" name="ifr2"></iframe>
<iframe src="./" name="ifr3"></iframe>

<script>
setup({'explicit_done':true})
var results = {};
var expected = {
    ifr1:{url:"about:blank", sameDom: true},
    ifr2:{url:"about:blank", sameDom: true},
    ifr3:{url: location.href.replace(/\/[^\/]*$/, '/'), sameDom: true },
    ifr4:{url:"about:blank", sameDom: true}
}

var js_url = 'javascript:"<html><script>var sameDom = false; try{var cn = top.document.body.className;sameDom = true;}catch(e){}; parent.postMessage( {url: document.URL, name: name, sameDom: sameDom}, \'*\')<\/script><body><p>JS-generated document</p></body></<html>";'
window.addEventListener('message', function(e){ console.log(e.data);
    var ifr = e.data.name;
    results[ifr] = e.data;
    test(function(){
        assert_equals(results[ifr].url, expected[ifr].url);
        assert_equals(results[ifr].sameDom, expected[ifr].sameDom);
    }, 'Testing URL and details of IFRAME ' + ifr);
    if(Object.keys(results).length === Object.keys(expected).length){
        done();
    }
}, false);

var ifr = document.createElement('iframe');
ifr.name = 'ifr4';
document.body.appendChild(ifr);

window.onload = function () {
    for (var i = 0, frame, frames = document.getElementsByTagName('iframe'); frame = frames[i]; i++) {
        try{
            console.log('setting src on ' + i);
            frame.src = js_url;
            console.log('done');
        }catch(e){
            console.log('exception')
            results[frame.name] = 'Exception on setting!';
            console.log('could read name')
        }
    };
}

</script>
</body>
</html>
